A* implementation attempt
This causes the pathfinding function in tileMap.cpp to fail quite often
when the path is a couple tiles away. It has to do with the way you
backtrack, since you search for the lowest distance... but that could be
the tile you were just at before... ending up in an infinite loop
essentially.