Temporary bug-fix for A* path finding bug
Put in a check to make sure the total path coordinates don't reach
numbers past ~300. This is a quick fix that at least stops it from
entering an infinite loop, but doesn't address the main issue. (the
problem stems from the way the openCoordinates are sorted by the
heuristic values).