NPC item collection working
The NPC is now able to keep track of the items it collects.

There are also changes that tweak its behavior so it seeks items, rather
than following the player. (there are still bugs present in the path
finding however so it gets stuck frequently).

The Leader class was implemented, which is for any player/npc characters
so they can track item collection or other "player-ish" needs. The
Leader class inherits from the Sprite class.

The command prompt prints the "score" between the player/npc.