Unfinished side menu and color scheme change
Added in the side menu class that currently has no real features. (just
a box that slides out from the right when pressing tab).

Also changed the color scheme of the game to a sort of retro 80's
computer chip look.

note: for some reason changing the colors for the tile resulted in the
program freezing when freeing one of the blitted surfaces. It's
currently commented out but is a memory leak to be fixed.