Player Inventory Grid Added
-Added in a vector that stores a grid of SDL_Rects to the Player class.
This serves as an outline for adding in the rest of the inventory
aspects.

NOTE: Using the grid of rects like this and having to update them
constantly seems temporary and sloppy. You should eventually make the
tileMap class (or sideMenu) more versatile so you can use that rather
than creating something new within the player class.